Refactor posture feature constants and types; streamline imports and remove unused code
