+++ src/features/posture/index.ts
@@ -1,4 +1,3 @@
-export { POSTURE_SPEC } from "./engine";
 export { usePostureTracking } from "./hooks/usePostureTracking";
 export { PostureControlPanel } from "./components/PostureControlPanel";
 export { PostureViewer } from "./components/PostureViewer";
