+++ src/features/posture/utils/angles.ts
@@ -1,0 +1,10 @@
+export function normalizeAngle(angle: number) {
+  let next = angle;
+  while (next > Math.PI) {
+    next -= Math.PI * 2;
+  }
+  while (next < -Math.PI) {
+    next += Math.PI * 2;
+  }
+  return next;
+}
